+++ os_computer_use/llm_provider.py
@@ -207,3 +207,98 @@ class AnthropicBaseProvider(LLMProvider):
         # Only return response text
         else:
             return text
+
+
+class MistralBaseProvider(OpenAIBaseProvider):
+    def create_client(self):
+        return OpenAI(
+            base_url=self.base_url,
+            api_key=self.api_key
+        ).chat.completions
+
+    def completion(self, messages, **kwargs):
+        try:
+            # Process message content
+            formatted_messages = []
+            for msg in messages:
+                content = msg["content"]
+                processed_content = []
+
+                if isinstance(content, list):
+                    for item in content:
+                        if isinstance(item, bytes):
+                            # Handle image data
+                            base64_image = base64.b64encode(item).decode('utf-8')
+                            processed_content.append({
+                                "type": "image_url",
+                                "image_url": {
+                                    "url": f"data:image/jpeg;base64,{base64_image}"
+                                }
+                            })
+                        else:
+                            # Handle text blocks
+                            processed_content.append({
+                                "type": "text",
+                                "text": str(item)
+                            })
+                else:
+                    # Handle simple text content
+                    processed_content.append({
+                        "type": "text",
+                        "text": str(content)
+                    })
+
+                formatted_messages.append({
+                    "role": msg["role"],
+                    "content": processed_content
+                })
+
+            # Prepare tools if provided
+            if "tools" in kwargs and kwargs["tools"] is not None:
+                kwargs["tools"] = [{
+                    "type": "function",
+                    "function": {
+                        "name": name,
+                        "description": details["description"],
+                        "parameters": {
+                            "type": "object",
+                            "properties": details["params"],
+                            "required": list(details["params"].keys())
+                        }
+                    }
+                } for name, details in kwargs["tools"].items()]
+
+            # Filter out None values
+            filtered_kwargs = {k: v for k, v in kwargs.items() if v is not None}
+            
+            return self.client.create(
+                model=self.model,
+                messages=formatted_messages,
+                **filtered_kwargs
+            )
+            
+        except Exception as e:
+            print(f"Mistral API Error: {str(e)}")
+            raise
+
+    def call(self, messages, functions=None):
+        try:
+            completion = self.completion(messages, tools=functions)
+            message = completion.choices[0].message
+            content = message.content
+            
+            # Extract tool calls
+            tool_calls = []
+            if message.tool_calls:
+                for tool_call in message.tool_calls:
+                    tool_calls.append({
+                        "type": "function",
+                        "name": tool_call.function.name,
+                        "parameters": json.loads(tool_call.function.arguments)
+                    })
+            
+            return content, tool_calls
+            
+        except json.JSONDecodeError:
+            print("Failed to parse tool call arguments")
+            return content, []
